misc: change the store & transport interface
1. the original store uses pub/sub pattern and it works not well on a
server cluster, so we want to change it, and use a message queue pattern
to instead it.

2. we remove some features the original socket.io has such as room,
namespace broadcast, because these features are a bit hard to support.
And in general, developpers would implements them at application layer.
However, if possible, we will re-implement them in the future.

3. to adapt the store interface, we refactor the implemetation of
transports, especially for long-polling.

4. we also remove the support of htmlfile. if possible, we may
re-implement it in the future.